pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sre-agent"
version = "0.1.0"
description = "SRE Multi-Agent System for automated Root Cause Analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents[anthropic]",
    "strands-agents-tools",
    "mcp[cli]",
    "fastmcp",
    "httpx",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml",
    "typer[all]",
    "rich",
]

[project.optional-dependencies]
ssh = ["paramiko"]
webhook = ["fastapi", "uvicorn"]
all = ["sre-agent[ssh,webhook]"]
dev = ["pytest", "pytest-asyncio", "ruff"]

[project.scripts]
sre-agent = "sre_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/sre_agent"]

[tool.ruff]
target-version = "py311"
line-length = 120
